Created Workload component.
